requirements.txt
beautifulsoup4==4.12.3
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
docopt==0.6.2
Flask==3.0.0
Flask-Cors==4.0.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
gspread==5.11.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pipreqs==0.4.13
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-decouple==3.8
pytweening==1.0.7
pywhatkit==5.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
urllib3==2.0.6
Werkzeug==3.0.0
wikipedia==1.4.0
yarg==0.1.9
python-dotenv
flask_cors
